Restore original `package-lock.json`
It looks like the `package-lock.json` file got reverted back to `lockfileVersion` 1, and then brought forward to v2, and in process some things got lost or munged.

This commit just copies the original `package-lock.json` into this branch in the hopes that this will fix the problems running the E2E tests in GitHub Actions.